Created functions
Created specific functions to create water/fertilizer notifications for a plant as well as function to cancel them
Also added some variables to the plant class to store their notification ID

+++ shared/src/commonMain/kotlin/com/gmg/growmygarden/NotificationHandler.kt
@@ -17,7 +17,7 @@ object NotificationHandler {
 
     private val localService = alarmeeService.local
 
-    fun setNotification(id: String, title: String, body: String, date: LocalDateTime, image: String?, delay: Int) {
+    fun setNotification(id: String, title: String, body: String, date: LocalDateTime, image: String?, delay: Long) {
         localService.schedule(
             alarmee = Alarmee(
                 uuid = id,
